Added Enpoint from ExpressPos todos los endpoints necesarios según la lista:
Generación de Reportes:

✅ Obtener Ventas por Día
✅ Obtener Detalle de Ventas
✅ Obtener Resumen de Cortes por Día
Gestión de Cortes de Caja:

✅ Registrar Corte
✅ Eliminar Corte
✅ Obtener Corte por ID
✅ Obtener Todos los Cortes
✅ Obtener Cortes por Rango
Gestión de Productos:

✅ Agregar Producto
✅ Actualizar Producto
✅ Eliminar Producto
✅ Obtener Producto por ID
✅ Obtener Todos los Productos
✅ Actualizar Stock
✅ Obtener Productos con Bajo Stock
Gestión de Ventas:

✅ Registrar Venta
✅ Eliminar Venta
✅ Obtener Venta por ID
✅ Obtener Ventas por Rango

+++ Medical.Office.App/Dtos/EspressPos/VentasDto.cs
@@ -4,5 +4,5 @@ public record VentasDto(
 
     int VentaID,
     DateTime? FechaHora,
-    decimal Total
+    double Total
 );
